+++ src/plugins/index.ts
@@ -2,55 +2,85 @@ import { Plugin } from '@nuxt/types'
 import { Storage } from './storage'
 
 const options = JSON.parse(`<%= JSON.stringify(options) %>`)
-const { registerUrl, loginUrl, userUrl, expires } = options
+const {
+  registerUrl,
+  loginUrl,
+  userUrl,
+  editUrl,
+  expires,
+  tokenProperty,
+  refreshTokenProperty,
+  tokenType,
+} = options
 const myPlugin: Plugin = (ctx, inject) => {
   const storage = new Storage(ctx)
   const allFn = {
-    register(data: any) {
-      return new Promise((resolve) => {
+    register(data: any): Promise<Object | void> {
+      return new Promise((resolve, reject) => {
         ctx.$axios
           .$post(registerUrl, data)
           .then((response) => resolve(response))
+          .catch((error) => {
+            reject(error)
+          })
       })
     },
-    async login({ username, password }: any): Promise<Object | void> {
-      const res = await ctx.$axios
-        .$post(loginUrl, { username, password })
-        .then((response) => {
-          return response
-        })
-        .catch((_) => {
-          this.logout()
-        })
-      const { access_token, refresh_token, token_type } = res
-      if (access_token) {
-        const expired = new Date().getTime() + expires * 1000
-        storage.setUniversal('token', access_token)
-        storage.setUniversal('refreshToken', refresh_token)
-        storage.setUniversal('expiredAt', expired)
-        storage.setUniversal('tokenType', token_type)
+    login({ username, password }: any): Promise<unknown> {
+      return new Promise((resolve, reject) => {
+        ctx.$axios
+          .$post(loginUrl, { username, password })
+          .then((response) => {
+            if (response[tokenProperty]) {
+              const expired = new Date().getTime() + expires * 1000
+              storage.setUniversal('token', response[tokenProperty])
+              storage.setUniversal(
+                'refreshToken',
+                response[refreshTokenProperty]
+              )
+              storage.setUniversal('expiredAt', expired)
+              storage.setUniversal('tokenType', tokenType)
 
+              resolve(this.fetchUserData())
+            } else {
+              resolve(response)
+            }
+          })
+          .catch((error) => {
+            reject(error)
+          })
+      })
+    },
+    editInfo(data: object): Promise<Object | void> {
+      return new Promise((resolve, reject) => {
         ctx.$axios.setToken(storage.state.token, storage.state.tokenType)
-        const userData = await ctx.$axios
+        ctx.$axios
+          .$put(editUrl, data)
+          .then((response) => {
+            resolve(response)
+          })
+          .catch((error) => {
+            reject(error)
+          })
+      })
+    },
+    fetchUserData(): Promise<unknown> {
+      return new Promise((resolve, reject) => {
+        ctx.$axios.setToken(storage.state.token, storage.state.tokenType)
+        ctx.$axios
           .$get(userUrl)
           .then((response) => {
-            return response
+            if (response) {
+              storage.setUniversal('user', response)
+              storage.setUniversal('isLogged', true)
+              resolve({ success: true })
+            } else {
+              resolve(response)
+            }
           })
-          .catch((_) => {
-            this.logout()
+          .catch((error) => {
+            reject(error)
           })
-        if (userData) {
-          storage.setUniversal('user', userData)
-          storage.setUniversal('isLogged', true)
-          return { success: true }
-        } else {
-          this.logout()
-          return userData
-        }
-      } else {
-        this.logout()
-        return res
-      }
+      })
     },
     logout(): Boolean | void {
       try {