+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - add option `guestPath`
 - remove vuex store replace with cookie
 - add function logout()
+- add function fetchUser()
+- add function editInfo()
 - add return value and error log in api refrence
 - add storage class
 - add userdata feature
@@ -11,6 +13,11 @@
 - add option `registerUrl`
 - add option `loginUrl`
 - add option `userUrl`
+- add option `editUrl`
+- add option `tokenProperty`
+- add option `refreshTokenProperty`
+- add option `tokenType`
+- add error handler
 
 ### 1.0.0